cleanup: Move many config to internal state translation functions into their own files.
